requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cssbeautifier==1.14.9
Django==4.2.7
django-admin-soft-dashboard==1.0.12
django-cpf==0.1.0
django-phonenumber-field==7.2.0
djlint==1.32.1
EditorConfig==0.12.3
gunicorn==21.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
jsbeautifier==1.14.9
json5==0.9.14
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
phonenumbers==8.13.24
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
sdist==0.0.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.5
whitenoise==6.5.0
